Change the timezone type used by ZonedDate and ZonedDateTime

+++ src/date/datetime.h
@@ -8,7 +8,7 @@
 typedef struct
 {
     TimestampTz value;
-    int16 zone;
+    int32 zone;
 } ZonedDateTime;
